Upgrades Play 2.3 -> 3.0

+++ app/formats/json/CommentSubmissionFormats.scala
@@ -1,7 +1,7 @@
 package formats.json
 
-import play.api.libs.json.{JsPath, Reads}
 import play.api.libs.functional.syntax._
+import play.api.libs.json.{JsPath, Reads}
 
 object CommentSubmissionFormats {
   case class CommentSubmission(auditTaskId: Int, missionId: Int, streetEdgeId: Int, comment: String,
@@ -10,11 +10,11 @@ object CommentSubmissionFormats {
 
   case class ValidationCommentSubmission(missionId: Int, labelId: Int, comment: String,
                                          gsvPanoramaId: String, heading: Double, pitch: Double,
-                                         zoom: Int, lat: Double, lng: Double)
+                                         zoom: Float, lat: Double, lng: Double)
 
-  case class LabelMapValidationCommentSubmission(labelId: Int, labelType: String, comment: String, 
-                                        gsvPanoramaId: String, heading: Double, pitch: Double, zoom: Int, 
-                                        lat: Double, lng: Double)
+  case class LabelMapValidationCommentSubmission(labelId: Int, labelType: String, comment: String,
+                                                 gsvPanoramaId: String, heading: Double, pitch: Double, zoom: Float,
+                                                 lat: Double, lng: Double)
 
   implicit val commentSubmissionReads: Reads[CommentSubmission] = (
     (JsPath \ "audit_task_id").read[Int] and
@@ -36,11 +36,11 @@ object CommentSubmissionFormats {
       (JsPath \ "gsv_panorama_id").read[String] and
       (JsPath \ "heading").read[Double] and
       (JsPath \ "pitch").read[Double] and
-      (JsPath \ "zoom").read[Int] and
+      (JsPath \ "zoom").read[Float] and
       (JsPath \ "lat").read[Double] and
       (JsPath \ "lng").read[Double]
   )(ValidationCommentSubmission.apply _)
-  
+
   implicit val labelMapValidationCommentSubmissionReads : Reads[LabelMapValidationCommentSubmission] = (
     (JsPath \ "label_id").read[Int] and
       (JsPath \ "label_type").read[String] and
@@ -48,7 +48,7 @@ object CommentSubmissionFormats {
       (JsPath \ "gsv_panorama_id").read[String] and
       (JsPath \ "heading").read[Double] and
       (JsPath \ "pitch").read[Double] and
-      (JsPath \ "zoom").read[Int] and
+      (JsPath \ "zoom").read[Float] and
       (JsPath \ "lat").read[Double] and
       (JsPath \ "lng").read[Double]
   )(LabelMapValidationCommentSubmission.apply _)
